+++ CHANGES.rst
@@ -7,3 +7,4 @@ Changelog
 
 - Has three abstract model classes for time stamped, user stamped and time + user stamped models.
 - Added an admin class to handle UserStampedModel and UserTimeStampedModel classes.
+- Added an admin inline classes to handle UserStampedModel and UserTimeStampedModel classes.